Spell effect restructure (#31565)
* move spell effects to their own file

* make function definitions clearer

removed namespace{} and put namespace at each definition.

* fix includes

* rename spell_effect.cpp to magic_spell_effect.cpp
